// frontend/src/lib/api/models/CreatePersonDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreatePersonDto = {
    firstName: string;
    lastName?: string | null;
    phone?: string | null;
    address?: string | null;
    /**
     * Fecha de nacimiento (ISO 8601)
     */
    birthDate?: string | null;
    isBaptizedMember?: boolean;
    notes?: string | null;
};